+++ changelog.md
@@ -5,6 +5,7 @@
 * LLT-5096: Fix uapi get/set to throw error when device dead
 * LLT-5017: Fix interface with static IP may be skipped for socket binding
 * LLT-5148: Rebind sockets when network path changes on Apple
+* LLT-5023: Add no-link enhanced detection mod
 
 <br>